% Generated by roxygen2 (4.0.0): do not edit by hand
\name{db.new}
\alias{db.new}
\title{Connection to a database}
\usage{
db.new(hostname = NA, port = NA, username = NA, password = NA,
  dbname = NA, dbtype = NA, profile = NA)
}
\arguments{
\item{hostname}{the hostname for your database (i.e. dw.muppets.com)}

\item{port}{the port for your database (i.e. 5432 or 3309)}

\item{username}{the username for your database (i.e. kermit)}

\item{password}{the password for your database (i.e. supersecret)}

\item{dbname}{the dbname for your database (i.e. dw, production, staging)}

\item{dbtype}{the dbtype for your database (postgres, mysql, sqlite, or redshift)}

\item{profile}{the profile for your database (a connection profile)}
}
\description{
Returns a structure that helps you query and interact with your database.
}
\examples{
\dontrun{
db <- db.new(username="kermit", password="rainbowconnection", hostname="localhost", dbname="muppetdb", dbtype="postgres")
db$save_credentails(profile="muppetdb")

db <- db.new(profile="mysql_local")
db$query("select * from foo limit 10;")
}
}
\keyword{db,}
\keyword{new}
\keyword{object,}

