% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DemoDB}
\alias{DemoDB}
\title{Demo database for testing and examples.}
\usage{
DemoDB()
}
\description{
Psuedo-function that returns an example database. We're using the Chinook
database which is a commonly used example database.
}
\examples{
db <- DemoDB()
db$query("select * from Artist;")
db$tables
}

