% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{queries.mysql}
\alias{queries.mysql}
\title{Queries for Mysql}
\format{\preformatted{List of 3
 $ column:List of 4
  ..$ head  : chr "select \%s from \%s limit \%d;"
  ..$ all   : chr "select \%s from \%s;"
  ..$ unique: chr "select distinct \%s from \%s;"
  ..$ sample: chr "select \%s from \%s order by rand() limit \%d;"
 $ table :List of 5
  ..$ select: chr "select \%s from \%s;"
  ..$ head  : chr "select * from \%s limit \%d;"
  ..$ all   : chr "select * from \%s;"
  ..$ unique: chr "select distinct \%s from \%s;"
  ..$ sample: chr "select * from \%s order by rand() limit \%d;"
 $ system:List of 6
  ..$ schema_no_system       : chr "                select                    table_name                    , column_name                    , data_type           "| __truncated__
  ..$ schema_with_system     : chr "                select                    table_name                    , column_name                    , data_type           "| __truncated__
  ..$ schema_specified       : chr "                select                    table_name                    , column_name                    , udt_name            "| __truncated__
  ..$ foreign_keys_for_table : chr "        select            column_name            , referenced_table_name            , referenced_column_name        from       "| __truncated__
  ..$ foreign_keys_for_column: chr "        select            column_name            , referenced_table_name            , referenced_column_name        from       "| __truncated__
  ..$ ref_keys_for_table     : chr "            select                referenced_column_name                , table_name                , column_name            fr"| __truncated__
}}
\usage{
queries.mysql
}
\description{
Queries for Mysql
}
\keyword{datasets}

