% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{queries.redshift}
\alias{queries.redshift}
\title{Queries for Redshift}
\format{\preformatted{List of 3
 $ column:List of 4
  ..$ head  : chr "select \%s from \%s limit \%d;"
  ..$ all   : chr "select \%s from \%s;"
  ..$ unique: chr "select distinct \%s from \%s;"
  ..$ sample: chr "select \%s from \%s order by random() limit \%d;"
 $ table :List of 5
  ..$ select: chr "select \%s from \%s;"
  ..$ head  : chr "select * from \%s limit \%d;"
  ..$ all   : chr "select * from \%s;"
  ..$ unique: chr "select distinct \%s from \%s;"
  ..$ sample: chr "select * from \%s order by random() limit \%d;"
 $ system:List of 5
  ..$ schema_no_system       : chr "                select                    table_name                    , column_name                    , udt_name            "| __truncated__
  ..$ schema_with_system     : chr "                select                    table_name                    , column_name                    , udt_name            "| __truncated__
  ..$ foreign_keys_for_table : chr "            SELECT                kcu.column_name                , ccu.table_name AS foreign_table_name                , ccu.co"| __truncated__
  ..$ foreign_keys_for_column: chr "            SELECT                kcu.column_name                , ccu.table_name AS foreign_table_name                , ccu.co"| __truncated__
  ..$ ref_keys_for_table     : chr "            SELECT                ccu.column_name                , kcu.table_name AS foreign_table_name                , kcu.co"| __truncated__
}}
\usage{
queries.redshift
}
\description{
Queries for Redshift
}
\keyword{datasets}

