% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{queries.sqlite}
\alias{queries.sqlite}
\title{Queries for SQLite}
\format{\preformatted{List of 3
 $ column:List of 4
  ..$ head  : chr "select \%s from \%s limit \%d;"
  ..$ all   : chr "select \%s from \%s;"
  ..$ unique: chr "select distinct \%s from \%s;"
  ..$ sample: chr "select \%s from \%s order by random() limit \%d;"
 $ table :List of 5
  ..$ select: chr "select \%s from \%s;"
  ..$ head  : chr "select * from \%s limit \%d;"
  ..$ all   : chr "select * from \%s;"
  ..$ unique: chr "select distinct \%s from \%s;"
  ..$ sample: chr "select * from \%s order by random() limit \%d;"
 $ system:List of 5
  ..$ schema_no_system       : chr "select table_name, column_name, data_type from tmp_dbpy_schema;"
  ..$ schema_with_system     : chr "select table_name, column_name, data_type from tmp_dbpy_schema;"
  ..$ foreign_keys_for_table : chr "            select                column_name                , foreign_table as foreign_table_name                , foreign_col"| __truncated__
  ..$ foreign_keys_for_column: chr "            select                column_name                , foreign_table as foreign_table_name                , foreign_col"| __truncated__
  ..$ ref_keys_for_table     : chr "            select                 foreign_column                 , table_name                 , column_name            from   "| __truncated__
}}
\usage{
queries.sqlite
}
\description{
Queries for SQLite
}
\keyword{datasets}

